# openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedResponse": {
        "description": "Standard envelope for list endpoints.",
        "properties": {
          "items": {
            "items": {},
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items"
        ],
        "title": "PaginatedResponse",
        "type": "object"
      },
...
      "PluginDiagnostics": {
        "properties": {
          "dependencies": {
            "items": {
              "type": "string"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "functionCount": {
            "default": 0,
            "title": "Functioncount",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "variableNames": {
            "items": {
              "type": "string"
            },
            "title": "Variablenames",
            "type": "array"
          },
          "workflowCount": {
            "default": 0,
            "title": "Workflowcount",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "title": "PluginDiagnostics",
        "type": "object"
      },
      "PluginFileWrite": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "expectedHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedhash"
          }
        },
        "required": [
          "content"
        ],
        "title": "PluginFileWrite",
        "type": "object"
      },
      "PluginManifestUpdate": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          }
        },
        "title": "PluginManifestUpdate",
        "type": "object"
      },
      "PluginResponse": {
        "properties": {
          "author": {
            "default": "",
            "title": "Author",
            "type": "string"
          },
          "category": {
            "default": "",
            "title": "Category",
            "type": "string"
          },
          "dependencies": {
            "items": {
              "type": "string"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "functionCount": {
            "default": 0,
            "title": "Functioncount",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "default": "discovered",
            "title": "Status",
            "type": "string"
          },
          "variableNames": {
            "items": {
              "type": "string"
            },
            "title": "Variablenames",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "workflowCount": {
            "default": 0,
            "title": "Workflowcount",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "version"
        ],
        "title": "PluginResponse",
        "type": "object"
      },
      "PluginScaffoldRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "PluginScaffoldRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first API workflow studio",
    "title": "SCLPLAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/plugins": {
      "get": {
        "description": "List all discovered plugins for a project.",
        "operationId": "list_plugins_api_v1_projects__project_id__plugins_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Plugins",
        "tags": [
          "plugins"
        ]
      },
      "post": {
        "description": "Create a new plugin scaffold.",
        "operationId": "scaffold_plugin_api_v1_projects__project_id__plugins_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginScaffoldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scaffold Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/reload": {
      "post": {
        "description": "Re-scan and reload all plugins.",
        "operationId": "reload_plugins_api_v1_projects__project_id__plugins_reload_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reload Plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}": {
      "get": {
        "description": "Get a single plugin with manifest and metadata.",
        "operationId": "get_plugin_api_v1_projects__project_id__plugins__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}/diagnostics": {
      "get": {
        "description": "Get plugin diagnostics with secrets excluded.\n\nReturns variable names but NOT values to protect secret content.",
        "operationId": "get_diagnostics_api_v1_projects__project_id__plugins__name__diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginDiagnostics"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Diagnostics",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}/disable": {
      "post": {
        "description": "Unload (disable) a plugin.",
        "operationId": "disable_plugin_api_v1_projects__project_id__plugins__name__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}/enable": {
      "post": {
        "description": "Load (enable) a plugin.",
        "operationId": "enable_plugin_api_v1_projects__project_id__plugins__name__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}/export": {
      "get": {
        "description": "Export a plugin as a zip archive.",
        "operationId": "export_plugin_api_v1_projects__project_id__plugins__name__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}/files/{path}": {
      "get": {
        "description": "Read a file within a plugin directory.",
        "operationId": "read_plugin_file_api_v1_projects__project_id__plugins__name__files__path__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Plugin File Api V1 Projects  Project Id  Plugins  Name  Files  Path  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Plugin File",
        "tags": [
          "plugins"
        ]
      },
      "put": {
        "description": "Write a file within a plugin directory.",
        "operationId": "write_plugin_file_api_v1_projects__project_id__plugins__name__files__path__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginFileWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Write Plugin File Api V1 Projects  Project Id  Plugins  Name  Files  Path  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Plugin File",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}/manifest": {
      "get": {
        "description": "Get the raw plugin.json manifest.",
        "operationId": "get_manifest_api_v1_projects__project_id__plugins__name__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Manifest Api V1 Projects  Project Id  Plugins  Name  Manifest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Manifest",
        "tags": [
          "plugins"
        ]
      },
      "patch": {
        "description": "Update the plugin.json manifest.",
        "operationId": "update_manifest_api_v1_projects__project_id__plugins__name__manifest_patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginManifestUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Manifest Api V1 Projects  Project Id  Plugins  Name  Manifest Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Manifest",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/projects/{project_id}/plugins/{name}/tree": {
      "get": {
        "description": "Get the file tree for a plugin.",
        "operationId": "get_plugin_tree_api_v1_projects__project_id__plugins__name__tree_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Plugin Tree Api V1 Projects  Project Id  Plugins  Name  Tree Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plugin Tree",
        "tags": [
          "plugins"
        ]
      }
    },
...
  }
}